\name{pathview-package}
\alias{pathview-package}
\docType{package}
\title{
  Pathway based data integration and visualization
}
\description{
  Pathway based data integration and visualization
}
\details{
\tabular{ll}{
Package: \tab pathview\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-26\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
  Weijun Luo <luo_weijun@yahoo.com>

Maintainer: Weijun Luo <luo_weijun@yahoo.com>
}
\references{
  Luo, W. and Brouwer, C., Pathview: an R/Bioconductor package for
  pathway based data integration and visualization. Bioinformatics,
 2013, 29(14): 1830-1831, doi: 10.1093/bioinformatics/btt285
}
\keyword{package}
